fixed baHistory repo and user repo

+++ src/main/java/com/example/leadManagementSystem2/Repository/BusinessAssociateHistoryRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.leadManagementSystem2.Entity.BusinessAssociateHistory;
 
-public interface BusinessAssociateHistoryRepo extends JpaRepository<BusinessAssociateHistory, Integer>{
+public interface BusinessAssociateHistoryRepo extends JpaRepository<BusinessAssociateHistory, Long>{
 
 }
